importlib + _imp + _ast + errno + _codecs + copyreg: extract into module/{importlib,_imp,_ast,errno,_codecs,copyreg}/
BUILTIN_MODULES thread-local promoted to pub(crate) so _imp.is_builtin
can still inspect the registry.  importing.rs shrinks by 695 lines.
