+++ pyre/pyre-interpreter/src/importing.rs
@@ -78,7 +78,7 @@ thread_local! {
     ///
     /// Maps module name → initializer function that populates a DictStorage.
     /// Each builtin module is lazily created on first import.
-    static BUILTIN_MODULES: RefCell<HashMap<&'static str, fn(&mut DictStorage)>> =
+    pub(crate) static BUILTIN_MODULES: RefCell<HashMap<&'static str, fn(&mut DictStorage)>> =
         RefCell::new(HashMap::new());
 }
 
@@ -123,21 +123,33 @@ pub fn install_builtin_modules() {
         "_contextvars",
         crate::module::_contextvars::moduledef::init,
     );
-    register_builtin_module("copyreg", init_copyreg);
-    register_builtin_module("_codecs", init_codecs);
+    register_builtin_module("copyreg", crate::module::copyreg::moduledef::init);
+    register_builtin_module("_codecs", crate::module::_codecs::moduledef::init);
     register_builtin_module("posix", crate::module::posix::moduledef::init);
-    register_builtin_module("errno", init_errno);
+    register_builtin_module("errno", crate::module::errno::moduledef::init);
     register_builtin_module(
         "_collections",
         crate::module::_collections::moduledef::init,
     );
-    register_builtin_module("_ast", init_ast);
+    register_builtin_module("_ast", crate::module::_ast::moduledef::init);
     register_builtin_module("_opcode", crate::module::_opcode::moduledef::init);
-    register_builtin_module("_imp", init_imp);
-    register_builtin_module("importlib.machinery", init_importlib_machinery);
-    register_builtin_module("importlib", init_importlib_pkg);
-    register_builtin_module("importlib.util", init_importlib_util);
-    register_builtin_module("importlib.abc", init_importlib_abc);
+    register_builtin_module("_imp", crate::module::_imp::moduledef::init);
+    register_builtin_module(
+        "importlib.machinery",
+        crate::module::importlib::moduledef::init_machinery,
+    );
+    register_builtin_module(
+        "importlib",
+        crate::module::importlib::moduledef::init_pkg,
+    );
+    register_builtin_module(
+        "importlib.util",
+        crate::module::importlib::moduledef::init_util,
+    );
+    register_builtin_module(
+        "importlib.abc",
+        crate::module::importlib::moduledef::init_abc,
+    );
     register_builtin_module("_signal", crate::module::_signal::moduledef::init);
     register_builtin_module("atexit", crate::module::atexit::moduledef::init);
     #[cfg(unix)]
@@ -237,687 +249,6 @@ fn init_sysconfigdata_empty(ns: &mut DictStorage) {
     // similar. Leave the dict empty; .get('X') returns None for unknown
     // keys which every caller already handles.
     crate::dict_storage_store(ns, "build_time_vars", vars);
-}
-
-/// importlib stub — PyPy: pypy/module/importlib/
-/// Avoid loading the real importlib.__init__ since it drags in
-/// _bootstrap and _bootstrap_external.
-fn init_importlib_pkg(ns: &mut DictStorage) {
-    // importlib.import_module(name, package=None) — return an imported
-    // module by name. PyPy: Lib/importlib/__init__.py import_module →
-    // _bootstrap._gcd_import. We defer to the interpreter's importhook
-    // since it handles both builtins and source modules.
-    crate::dict_storage_store(
-        ns,
-        "import_module",
-        crate::make_builtin_function("import_module", |args| {
-            let name = args.first().copied().unwrap_or(pyre_object::w_none());
-            unsafe {
-                if !pyre_object::is_str(name) {
-                    return Err(crate::PyError::type_error(
-                        "import_module: name must be str",
-                    ));
-                }
-                let name_str = pyre_object::w_str_get_value(name).to_string();
-                crate::importing::importhook(
-                    &name_str,
-                    pyre_object::w_none(),
-                    pyre_object::w_list_new(vec![pyre_object::w_str_new("*")]),
-                    0,
-                    std::ptr::null(),
-                )
-            }
-        }),
-    );
-    crate::dict_storage_store(
-        ns,
-        "invalidate_caches",
-        crate::make_builtin_function_with_arity(
-            "invalidate_caches",
-            |_| Ok(pyre_object::w_none()),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "reload",
-        crate::make_builtin_function_with_arity(
-            "reload",
-            |args| Ok(args.first().copied().unwrap_or(pyre_object::w_none())),
-            1,
-        ),
-    );
-    // Mark as a package so dotted imports treat it as such.
-    crate::dict_storage_store(ns, "__path__", pyre_object::w_list_new(vec![]));
-}
-
-/// importlib.util stub — minimal subset.
-fn init_importlib_util(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "spec_from_file_location",
-        crate::make_builtin_function_with_arity(
-            "spec_from_file_location",
-            |_| Ok(pyre_object::w_none()),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "module_from_spec",
-        crate::make_builtin_function_with_arity(
-            "module_from_spec",
-            |_| Ok(pyre_object::w_none()),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "find_spec",
-        crate::make_builtin_function_with_arity("find_spec", |_| Ok(pyre_object::w_none()), 0),
-    );
-    crate::dict_storage_store(
-        ns,
-        "resolve_name",
-        crate::make_builtin_function("resolve_name", |args| {
-            Ok(args.first().copied().unwrap_or(pyre_object::w_str_new("")))
-        }),
-    );
-    crate::dict_storage_store(ns, "MAGIC_NUMBER", pyre_object::w_int_new(0));
-}
-
-/// importlib.abc stub — abstract base classes.
-fn init_importlib_abc(ns: &mut DictStorage) {
-    for name in [
-        "Loader",
-        "Finder",
-        "MetaPathFinder",
-        "PathEntryFinder",
-        "ResourceLoader",
-        "InspectLoader",
-        "ExecutionLoader",
-        "FileLoader",
-        "SourceLoader",
-    ] {
-        crate::dict_storage_store(ns, name, crate::typedef::w_object());
-    }
-}
-
-/// importlib.machinery stub — provides the names inspect.py references.
-/// PyPy ships the real importlib; we shortcut it with a stub so pyre does
-/// not have to execute _bootstrap_external.
-fn init_importlib_machinery(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "SOURCE_SUFFIXES",
-        pyre_object::w_list_new(vec![pyre_object::w_str_new(".py")]),
-    );
-    crate::dict_storage_store(
-        ns,
-        "BYTECODE_SUFFIXES",
-        pyre_object::w_list_new(vec![pyre_object::w_str_new(".pyc")]),
-    );
-    crate::dict_storage_store(
-        ns,
-        "EXTENSION_SUFFIXES",
-        pyre_object::w_list_new(vec![pyre_object::w_str_new(".so")]),
-    );
-    crate::dict_storage_store(
-        ns,
-        "DEBUG_BYTECODE_SUFFIXES",
-        pyre_object::w_list_new(vec![pyre_object::w_str_new(".pyc")]),
-    );
-    crate::dict_storage_store(
-        ns,
-        "OPTIMIZED_BYTECODE_SUFFIXES",
-        pyre_object::w_list_new(vec![pyre_object::w_str_new(".pyc")]),
-    );
-    crate::dict_storage_store(
-        ns,
-        "all_suffixes",
-        crate::make_builtin_function_with_arity(
-            "all_suffixes",
-            |_| {
-                Ok(pyre_object::w_list_new(vec![
-                    pyre_object::w_str_new(".py"),
-                    pyre_object::w_str_new(".pyc"),
-                    pyre_object::w_str_new(".so"),
-                ]))
-            },
-            0,
-        ),
-    );
-    crate::dict_storage_store(ns, "ModuleSpec", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "BuiltinImporter", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "FrozenImporter", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "PathFinder", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "FileFinder", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "SourceFileLoader", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "SourcelessFileLoader", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "ExtensionFileLoader", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "AppleFrameworkLoader", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "NamespaceLoader", crate::typedef::w_object());
-    crate::dict_storage_store(ns, "WindowsRegistryFinder", crate::typedef::w_object());
-}
-
-/// _imp stub — PyPy: pypy/module/imp/
-///
-/// Minimal subset required by importlib._bootstrap to decide which loader
-/// handles a name. We report every name we know about as a builtin so
-/// pyre's own registrations remain authoritative.
-fn init_imp(ns: &mut DictStorage) {
-    crate::dict_storage_store(
-        ns,
-        "is_builtin",
-        crate::make_builtin_function_with_arity(
-            "is_builtin",
-            |args| {
-                if args.is_empty() {
-                    return Ok(pyre_object::w_int_new(0));
-                }
-                let name = unsafe {
-                    if pyre_object::is_str(args[0]) {
-                        pyre_object::w_str_get_value(args[0])
-                    } else {
-                        return Ok(pyre_object::w_int_new(0));
-                    }
-                };
-                let is_builtin = BUILTIN_MODULES.with(|m| m.borrow().contains_key(name));
-                Ok(pyre_object::w_int_new(if is_builtin { 1 } else { 0 }))
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "is_frozen",
-        crate::make_builtin_function_with_arity(
-            "is_frozen",
-            |_| Ok(pyre_object::w_bool_from(false)),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "is_frozen_package",
-        crate::make_builtin_function_with_arity(
-            "is_frozen_package",
-            |_| Ok(pyre_object::w_bool_from(false)),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "get_frozen_object",
-        crate::make_builtin_function_with_arity(
-            "get_frozen_object",
-            |_| Ok(pyre_object::w_none()),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "create_builtin",
-        crate::make_builtin_function_with_arity(
-            "create_builtin",
-            |args| {
-                if args.is_empty() {
-                    return Ok(pyre_object::w_none());
-                }
-                Ok(args[0])
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "exec_builtin",
-        crate::make_builtin_function_with_arity(
-            "exec_builtin",
-            |_| Ok(pyre_object::w_int_new(0)),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "exec_dynamic",
-        crate::make_builtin_function_with_arity(
-            "exec_dynamic",
-            |_| Ok(pyre_object::w_int_new(0)),
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "acquire_lock",
-        crate::make_builtin_function_with_arity("acquire_lock", |_| Ok(pyre_object::w_none()), 0),
-    );
-    crate::dict_storage_store(
-        ns,
-        "release_lock",
-        crate::make_builtin_function_with_arity("release_lock", |_| Ok(pyre_object::w_none()), 0),
-    );
-    crate::dict_storage_store(
-        ns,
-        "lock_held",
-        crate::make_builtin_function_with_arity(
-            "lock_held",
-            |_| Ok(pyre_object::w_bool_from(false)),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "_fix_co_filename",
-        crate::make_builtin_function_with_arity(
-            "_fix_co_filename",
-            |_| Ok(pyre_object::w_none()),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "extension_suffixes",
-        crate::make_builtin_function_with_arity(
-            "extension_suffixes",
-            |_| Ok(pyre_object::w_list_new(vec![])),
-            0,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "source_hash",
-        crate::make_builtin_function_with_arity(
-            "source_hash",
-            |_| Ok(pyre_object::w_int_new(0)),
-            2,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "check_hash_based_pycs",
-        pyre_object::w_str_new("default"),
-    );
-    crate::dict_storage_store(ns, "pyc_magic_number_token", pyre_object::w_int_new(3495));
-}
-
-/// _ast stub — PyPy: pypy/module/_ast/
-///
-/// Exposes the AST node type hierarchy as plain type stubs. Our stubs are
-/// enough to satisfy `from _ast import *` in `ast.py` and class body
-/// references like `class slice(AST)`. Actual AST construction is not
-/// supported because pyre uses RustPython's compiler.
-fn init_ast(ns: &mut DictStorage) {
-    let ast_names: &[&str] = &[
-        "AST",
-        "mod",
-        "Module",
-        "Interactive",
-        "Expression",
-        "FunctionType",
-        "stmt",
-        "FunctionDef",
-        "AsyncFunctionDef",
-        "ClassDef",
-        "Return",
-        "Delete",
-        "Assign",
-        "TypeAlias",
-        "AugAssign",
-        "AnnAssign",
-        "For",
-        "AsyncFor",
-        "While",
-        "If",
-        "With",
-        "AsyncWith",
-        "Match",
-        "Raise",
-        "Try",
-        "TryStar",
-        "Assert",
-        "Import",
-        "ImportFrom",
-        "Global",
-        "Nonlocal",
-        "Expr",
-        "Pass",
-        "Break",
-        "Continue",
-        "expr",
-        "BoolOp",
-        "NamedExpr",
-        "BinOp",
-        "UnaryOp",
-        "Lambda",
-        "IfExp",
-        "Dict",
-        "Set",
-        "ListComp",
-        "SetComp",
-        "DictComp",
-        "GeneratorExp",
-        "Await",
-        "Yield",
-        "YieldFrom",
-        "Compare",
-        "Call",
-        "FormattedValue",
-        "JoinedStr",
-        "Constant",
-        "Attribute",
-        "Subscript",
-        "Starred",
-        "Name",
-        "List",
-        "Tuple",
-        "Slice",
-        "expr_context",
-        "Load",
-        "Store",
-        "Del",
-        "boolop",
-        "And",
-        "Or",
-        "operator",
-        "Add",
-        "Sub",
-        "Mult",
-        "MatMult",
-        "Div",
-        "Mod",
-        "Pow",
-        "LShift",
-        "RShift",
-        "BitOr",
-        "BitXor",
-        "BitAnd",
-        "FloorDiv",
-        "unaryop",
-        "Invert",
-        "Not",
-        "UAdd",
-        "USub",
-        "cmpop",
-        "Eq",
-        "NotEq",
-        "Lt",
-        "LtE",
-        "Gt",
-        "GtE",
-        "Is",
-        "IsNot",
-        "In",
-        "NotIn",
-        "comprehension",
-        "excepthandler",
-        "ExceptHandler",
-        "arguments",
-        "arg",
-        "keyword",
-        "alias",
-        "withitem",
-        "match_case",
-        "pattern",
-        "MatchValue",
-        "MatchSingleton",
-        "MatchSequence",
-        "MatchMapping",
-        "MatchClass",
-        "MatchStar",
-        "MatchAs",
-        "MatchOr",
-        "type_ignore",
-        "TypeIgnore",
-        "type_param",
-        "TypeVar",
-        "ParamSpec",
-        "TypeVarTuple",
-        // Flags used by ast.parse()
-        "PyCF_ONLY_AST",
-        "PyCF_OPTIMIZED_AST",
-        "PyCF_TYPE_COMMENTS",
-        "PyCF_ALLOW_TOP_LEVEL_AWAIT",
-    ];
-    for name in ast_names {
-        if name.starts_with("PyCF") {
-            crate::dict_storage_store(ns, name, pyre_object::w_int_new(0));
-        } else {
-            crate::dict_storage_store(ns, name, crate::typedef::make_builtin_type(name, |_| {}));
-        }
-    }
-}
-
-/// errno stub — PyPy: pypy/module/errno/
-fn init_errno(ns: &mut DictStorage) {
-    // Numerics differ per OS (e.g. EAGAIN is 11 on Linux but 35 on macOS),
-    // so when `host_env` is enabled we resolve every constant through
-    // `rustpython_host_env::errno::errors` (a `pub use libc::*` re-export).
-    // The `host_env = off` build keeps a darwin/BSD-flavoured fallback so
-    // pyre-wasm preserves its previous behaviour.
-    #[cfg(feature = "host_env")]
-    {
-        use rustpython_host_env::errno::errors as host_errno;
-        let entries: &[(&str, i32)] = &[
-            ("EPERM", host_errno::EPERM),
-            ("ENOENT", host_errno::ENOENT),
-            ("ESRCH", host_errno::ESRCH),
-            ("EINTR", host_errno::EINTR),
-            ("EIO", host_errno::EIO),
-            ("ENXIO", host_errno::ENXIO),
-            ("E2BIG", host_errno::E2BIG),
-            ("ENOEXEC", host_errno::ENOEXEC),
-            ("EBADF", host_errno::EBADF),
-            ("ECHILD", host_errno::ECHILD),
-            ("EAGAIN", host_errno::EAGAIN),
-            ("EWOULDBLOCK", host_errno::EWOULDBLOCK),
-            ("ENOMEM", host_errno::ENOMEM),
-            ("EACCES", host_errno::EACCES),
-            ("EFAULT", host_errno::EFAULT),
-            ("EBUSY", host_errno::EBUSY),
-            ("EEXIST", host_errno::EEXIST),
-            ("EXDEV", host_errno::EXDEV),
-            ("ENODEV", host_errno::ENODEV),
-            ("ENOTDIR", host_errno::ENOTDIR),
-            ("EISDIR", host_errno::EISDIR),
-            ("EINVAL", host_errno::EINVAL),
-            ("ENFILE", host_errno::ENFILE),
-            ("EMFILE", host_errno::EMFILE),
-            ("ENOTTY", host_errno::ENOTTY),
-            ("EFBIG", host_errno::EFBIG),
-            ("ENOSPC", host_errno::ENOSPC),
-            ("ESPIPE", host_errno::ESPIPE),
-            ("EROFS", host_errno::EROFS),
-            ("EMLINK", host_errno::EMLINK),
-            ("EPIPE", host_errno::EPIPE),
-            ("EDOM", host_errno::EDOM),
-            ("ERANGE", host_errno::ERANGE),
-            ("EDEADLK", host_errno::EDEADLK),
-            ("ENAMETOOLONG", host_errno::ENAMETOOLONG),
-            ("ENOLCK", host_errno::ENOLCK),
-            ("ENOSYS", host_errno::ENOSYS),
-            ("ENOTEMPTY", host_errno::ENOTEMPTY),
-            ("ELOOP", host_errno::ELOOP),
-            ("EOVERFLOW", host_errno::EOVERFLOW),
-            ("EPROTO", host_errno::EPROTO),
-            ("EDESTADDRREQ", host_errno::EDESTADDRREQ),
-            ("EAFNOSUPPORT", host_errno::EAFNOSUPPORT),
-            ("EALREADY", host_errno::EALREADY),
-            ("EDQUOT", host_errno::EDQUOT),
-        ];
-        for (name, value) in entries {
-            crate::dict_storage_store(ns, name, pyre_object::w_int_new(*value as i64));
-        }
-        // Unix-only constants (windows libc lacks some of these).
-        #[cfg(unix)]
-        {
-            let unix_entries: &[(&str, i32)] = &[
-                ("ENOTBLK", host_errno::ENOTBLK),
-                ("ETXTBSY", host_errno::ETXTBSY),
-                ("ENOMSG", host_errno::ENOMSG),
-                ("EIDRM", host_errno::EIDRM),
-                ("EBADMSG", host_errno::EBADMSG),
-                ("EMULTIHOP", host_errno::EMULTIHOP),
-                ("ENODATA", host_errno::ENODATA),
-                ("ENOLINK", host_errno::ENOLINK),
-                ("ENOSR", host_errno::ENOSR),
-                ("ENOSTR", host_errno::ENOSTR),
-                ("ETIME", host_errno::ETIME),
-            ];
-            for (name, value) in unix_entries {
-                crate::dict_storage_store(ns, name, pyre_object::w_int_new(*value as i64));
-            }
-        }
-    }
-    #[cfg(not(feature = "host_env"))]
-    {
-        let entries: &[(&str, i64)] = &[
-            ("EPERM", 1),
-            ("ENOENT", 2),
-            ("ESRCH", 3),
-            ("EINTR", 4),
-            ("EIO", 5),
-            ("ENXIO", 6),
-            ("E2BIG", 7),
-            ("ENOEXEC", 8),
-            ("EBADF", 9),
-            ("ECHILD", 10),
-            ("EAGAIN", 35),
-            ("EWOULDBLOCK", 35),
-            ("ENOMEM", 12),
-            ("EACCES", 13),
-            ("EFAULT", 14),
-            ("ENOTBLK", 15),
-            ("EBUSY", 16),
-            ("EEXIST", 17),
-            ("EXDEV", 18),
-            ("ENODEV", 19),
-            ("ENOTDIR", 20),
-            ("EISDIR", 21),
-            ("EINVAL", 22),
-            ("ENFILE", 23),
-            ("EMFILE", 24),
-            ("ENOTTY", 25),
-            ("ETXTBSY", 26),
-            ("EFBIG", 27),
-            ("ENOSPC", 28),
-            ("ESPIPE", 29),
-            ("EROFS", 30),
-            ("EMLINK", 31),
-            ("EPIPE", 32),
-            ("EDOM", 33),
-            ("ERANGE", 34),
-            ("EDEADLK", 11),
-            ("ENAMETOOLONG", 63),
-            ("ENOLCK", 77),
-            ("ENOSYS", 78),
-            ("ENOTEMPTY", 66),
-            ("ELOOP", 62),
-            ("ENOMSG", 91),
-            ("EIDRM", 90),
-            ("EBADMSG", 94),
-            ("EMULTIHOP", 95),
-            ("ENODATA", 96),
-            ("ENOLINK", 97),
-            ("ENOSR", 98),
-            ("ENOSTR", 99),
-            ("EOVERFLOW", 84),
-            ("EPROTO", 100),
-            ("ETIME", 101),
-            ("EDESTADDRREQ", 39),
-            ("EAFNOSUPPORT", 47),
-            ("EALREADY", 37),
-            ("EDQUOT", 69),
-        ];
-        for (name, value) in entries {
-            crate::dict_storage_store(ns, name, pyre_object::w_int_new(*value));
-        }
-    }
-    crate::dict_storage_store(ns, "errorcode", pyre_object::w_dict_new());
-}
-
-/// _codecs stub — PyPy: pypy/module/_codecs/
-///
-/// Provides lookup_error/register_error and encode/decode no-op stubs so
-/// codecs.py module init runs to completion.
-fn init_codecs(ns: &mut DictStorage) {
-    // lookup_error(name) — returns an error handler for the given error
-    // strategy. Pyre returns a pass-through lambda that never fires because
-    // we don't encounter encoding errors in the pure-Python stdlib paths
-    // we exercise so far.
-    crate::dict_storage_store(
-        ns,
-        "lookup_error",
-        crate::make_builtin_function_with_arity(
-            "lookup_error",
-            |_| {
-                Ok(crate::make_builtin_function_with_arity(
-                    "error_handler",
-                    |args| {
-                        Ok(if args.is_empty() {
-                            pyre_object::w_none()
-                        } else {
-                            args[0]
-                        })
-                    },
-                    1,
-                ))
-            },
-            1,
-        ),
-    );
-    crate::dict_storage_store(
-        ns,
-        "register_error",
-        crate::make_builtin_function_with_arity("register_error", |_| Ok(pyre_object::w_none()), 2),
-    );
-    crate::dict_storage_store(
-        ns,
-        "register",
-        crate::make_builtin_function_with_arity("register", |_| Ok(pyre_object::w_none()), 1),
-    );
-    crate::dict_storage_store(
-        ns,
-        "lookup",
-        crate::make_builtin_function_with_arity("lookup", |_| Ok(pyre_object::w_none()), 1),
-    );
-    // encode/decode — return input unchanged. Matches PyPy _codecs.encode
-    // when the codec is the identity.
-    let identity = crate::make_builtin_function_with_arity(
-        "identity",
-        |args| {
-            Ok(if args.is_empty() {
-                pyre_object::w_none()
-            } else {
-                args[0]
-            })
-        },
-        1,
-    );
-    crate::dict_storage_store(ns, "encode", identity);
-    crate::dict_storage_store(ns, "decode", identity);
-    crate::dict_storage_store(ns, "_forget_codec", identity);
-    crate::dict_storage_store(
-        ns,
-        "charmap_build",
-        crate::make_builtin_function_with_arity(
-            "charmap_build",
-            |_| Ok(pyre_object::w_dict_new()),
-            1,
-        ),
-    );
-}
-
-/// copyreg stub — PyPy: pypy/module/copyreg/
-fn init_copyreg(ns: &mut DictStorage) {
-    // copyreg.pickle(type, reduce_func, constructor=None) — register a
-    // pickle reducer. Stub: ignore (pyre doesn't support pickle).
-    crate::dict_storage_store(
-        ns,
-        "pickle",
-        crate::make_builtin_function_with_arity("pickle", |_| Ok(pyre_object::w_none()), 3),
-    );
-    crate::dict_storage_store(ns, "dispatch_table", pyre_object::w_dict_new());
 }
 
 /// Try to load a builtin module by name.
