+++ pyre/pyre-interpreter/src/module/_ast/interp_ast.rs
@@ -1,0 +1,152 @@
+//! _ast implementation — PyPy: pypy/module/_ast/interp_ast.py
+//!
+//! Verbatim move of the inline block previously in importing.rs.
+
+use crate::DictStorage;
+
+/// _ast stub — PyPy: pypy/module/_ast/
+///
+/// Exposes the AST node type hierarchy as plain type stubs. Our stubs are
+/// enough to satisfy `from _ast import *` in `ast.py` and class body
+/// references like `class slice(AST)`. Actual AST construction is not
+/// supported because pyre uses RustPython's compiler.
+pub fn register_module(ns: &mut DictStorage) {
+    let ast_names: &[&str] = &[
+        "AST",
+        "mod",
+        "Module",
+        "Interactive",
+        "Expression",
+        "FunctionType",
+        "stmt",
+        "FunctionDef",
+        "AsyncFunctionDef",
+        "ClassDef",
+        "Return",
+        "Delete",
+        "Assign",
+        "TypeAlias",
+        "AugAssign",
+        "AnnAssign",
+        "For",
+        "AsyncFor",
+        "While",
+        "If",
+        "With",
+        "AsyncWith",
+        "Match",
+        "Raise",
+        "Try",
+        "TryStar",
+        "Assert",
+        "Import",
+        "ImportFrom",
+        "Global",
+        "Nonlocal",
+        "Expr",
+        "Pass",
+        "Break",
+        "Continue",
+        "expr",
+        "BoolOp",
+        "NamedExpr",
+        "BinOp",
+        "UnaryOp",
+        "Lambda",
+        "IfExp",
+        "Dict",
+        "Set",
+        "ListComp",
+        "SetComp",
+        "DictComp",
+        "GeneratorExp",
+        "Await",
+        "Yield",
+        "YieldFrom",
+        "Compare",
+        "Call",
+        "FormattedValue",
+        "JoinedStr",
+        "Constant",
+        "Attribute",
+        "Subscript",
+        "Starred",
+        "Name",
+        "List",
+        "Tuple",
+        "Slice",
+        "expr_context",
+        "Load",
+        "Store",
+        "Del",
+        "boolop",
+        "And",
+        "Or",
+        "operator",
+        "Add",
+        "Sub",
+        "Mult",
+        "MatMult",
+        "Div",
+        "Mod",
+        "Pow",
+        "LShift",
+        "RShift",
+        "BitOr",
+        "BitXor",
+        "BitAnd",
+        "FloorDiv",
+        "unaryop",
+        "Invert",
+        "Not",
+        "UAdd",
+        "USub",
+        "cmpop",
+        "Eq",
+        "NotEq",
+        "Lt",
+        "LtE",
+        "Gt",
+        "GtE",
+        "Is",
+        "IsNot",
+        "In",
+        "NotIn",
+        "comprehension",
+        "excepthandler",
+        "ExceptHandler",
+        "arguments",
+        "arg",
+        "keyword",
+        "alias",
+        "withitem",
+        "match_case",
+        "pattern",
+        "MatchValue",
+        "MatchSingleton",
+        "MatchSequence",
+        "MatchMapping",
+        "MatchClass",
+        "MatchStar",
+        "MatchAs",
+        "MatchOr",
+        "type_ignore",
+        "TypeIgnore",
+        "type_param",
+        "TypeVar",
+        "ParamSpec",
+        "TypeVarTuple",
+        // Flags used by ast.parse()
+        "PyCF_ONLY_AST",
+        "PyCF_OPTIMIZED_AST",
+        "PyCF_TYPE_COMMENTS",
+        "PyCF_ALLOW_TOP_LEVEL_AWAIT",
+    ];
+    for name in ast_names {
+        if name.starts_with("PyCF") {
+            crate::dict_storage_store(ns, name, pyre_object::w_int_new(0));
+        } else {
+            crate::dict_storage_store(ns, name, crate::typedef::make_builtin_type(name, |_| {}));
+        }
+    }
+}
