+++ pyre/pyre-interpreter/src/module/_ast/moduledef.rs
@@ -1,0 +1,7 @@
+//! _ast module definition — PyPy: pypy/module/_ast/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_ast::register_module(ns);
+}
