+++ pyre/pyre-interpreter/src/module/_codecs/mod.rs
@@ -1,0 +1,7 @@
+//! _codecs module — PyPy: pypy/module/_codecs/
+//!
+//! Stub providing lookup_error / register_error and encode/decode no-op
+//! shells — enough for codecs.py module init to complete.
+
+pub mod interp_codecs;
+pub mod moduledef;
