+++ pyre/pyre-interpreter/src/module/copyreg/mod.rs
@@ -1,0 +1,4 @@
+//! copyreg module — PyPy: pypy/module/copyreg/
+
+pub mod interp_copyreg;
+pub mod moduledef;
