+++ pyre/pyre-interpreter/src/module/errno/mod.rs
@@ -1,0 +1,4 @@
+//! errno module — PyPy: pypy/module/errno/
+
+pub mod interp_errno;
+pub mod moduledef;
