+++ pyre/pyre-interpreter/src/module/importlib/mod.rs
@@ -1,0 +1,6 @@
+//! importlib module package — PyPy: Lib/importlib/ (app-level).  Pyre
+//! exposes a small builtin shim covering the four submodule namespaces
+//! `importlib`, `importlib.util`, `importlib.abc`, `importlib.machinery`.
+
+pub mod interp_importlib;
+pub mod moduledef;
